requirements.txt
Django==3.2.12
GPSPhoto==2.2.3
mysql_connector_repackaged==0.3.1
phonenumbers==8.12.45
Pillow==9.1.0
protobuf==3.20.0
pycryptodome==3.14.1
pyjks==20.0.0
reverse_geocoder==1.5.1
twilio==7.7.1
djangorestframework==3.13.1
mysqlclient==2.1.0
google-cloud-vision==2.7.1
exifread==2.3.2
piexif==1.1.3
docker-compose==1.29.2
pymemcache==3.5.1
pytest-shutil==1.7.0
django-extensions==3.1.5
gunicorn==20.1.0
django-widget-tweaks==1.4.12
